Add new Fields to `Service` Struct
Add `UpdateNotify` Field to allow the Notifiers to check if
`UpdateNotify` has been enabled in `CoreApp`.

Add `DownText` Field to have a default Downtime Text which also has the
'Service has been down for n-Seconds' Text.

Add `SuccessNotified` Field to check if a "Success" Message has already
been send.
This is because the `s.Online` Field is always "false" if the
OnSuccess() Function is called of an Notifier.