Fix email template
The Email notifier was using `map[string]interface{}` to "put" the
information (like Service and Failure struct) into the template.
The value of a map can only be used with `{{ index . "Service" }}` in a
GoLang template.
To simplify – and reduce the needed changes – `emailOutgoing` now
contains for the `Data` field an extra struct with `Service` and
`Failure` fields.