Permit the healthcheck to work if a BASE_PATH is set
You can set the environment variable BASE_PATH=statping so that https://yoursite.domain:1234/statping/ is how the app is accessed.  Note that a forward-slash is not used (which would otherwise allow just the variable to be dropped in). In order to do that, the healthcheck, which does not obey BASE_PATH, must be disabled. This should make it continue to work.